extract peelonion into the sphinx class
